Move Navigator to navigation package, remove interface
